chore: restructure 01 amend/fixup task to modify add()
With moving the order of the tasks I needed to not use the `ZeroDivsion` example as it wouldn't exist in the code
base. Instead I've removed the `Examples` section from the docstring of the `add()` function and that is added as a task
to test `amend` and then another example is added for testing `fixup`.